Added Memory GRPC Plugin. Renamed packages (#422)
Renamed packages from "external" to "grpc".
Changed storage type to "grpc-plugin".
Added cmd/memory-grpc-plugin that uses the memory backend as a grpc plugin.
Added go-plugin dependency to Glide.
Added task to generate plugin protos to Makefile.
