+++ plugin/storage/grpc/plugin.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package external
+package grpc
 
 import (
 	"context"
@@ -23,8 +23,8 @@ import (
 	"github.com/hashicorp/go-plugin"
 
 	"github.com/jaegertracing/jaeger/model"
-	"github.com/jaegertracing/jaeger/pkg/external/config"
-	"github.com/jaegertracing/jaeger/plugin/storage/external/shared"
+	"github.com/jaegertracing/jaeger/pkg/grpc/config"
+	"github.com/jaegertracing/jaeger/plugin/storage/grpc/shared"
 	"github.com/jaegertracing/jaeger/storage/spanstore"
 )
